OTAT-212 Fix bugs and add edit button to compare result page
