+++ frontend/src/components/compare/CompareResultContainer.tsx
@@ -1,13 +1,15 @@
 import React from "react";
 import Box from "@mui/material/Box";
 import { Trans } from "react-i18next";
-import { useParams, useSearchParams } from "react-router-dom";
+import { useNavigate, useLocation, useSearchParams } from "react-router-dom";
 import { useServiceContext } from "../../providers/ServiceProvider";
 import { useQuery } from "../../utils/useQuery";
 import QueryData from "../shared/QueryData";
 import Title from "../shared/Title";
 import CompareResult from "./CompareResult";
 import { ICompareResultModel } from "../../types";
+import Button from "@mui/material/Button";
+import BorderColorRoundedIcon from "@mui/icons-material/BorderColorRounded";
 
 const CompareResultContainer = () => {
   const [searchParams] = useSearchParams();
@@ -18,6 +20,8 @@ const CompareResultContainer = () => {
       service.fetchCompareResult(args, config),
     toastError: true,
   });
+  const navigate = useNavigate();
+  const location = useLocation();
 
   return (
     <QueryData
@@ -25,13 +29,26 @@ const CompareResultContainer = () => {
       render={(data) => {
         return (
           <Box>
-            <Title borderBottom={true}>
+            <Title
+              borderBottom={true}
+              toolbar={
+                <Button
+                  startIcon={<BorderColorRoundedIcon />}
+                  size="small"
+                  onClick={() =>
+                    navigate({ pathname: "/compare", search: location.search })
+                  }
+                >
+                  <Trans i18nKey="editComparisonItems" />
+                </Button>
+              }
+            >
               <Trans i18nKey="theResultOfComparing" />{" "}
               <Box sx={{ ml: { xs: 0, sm: 1 } }} display="inline">
                 {data.base_infos.map((item: any, index: number) => (
                   <>
                     {item.title}
-                    {index < data.base_infos.length - 1 ? " and" : ""}{" "}
+                    {getConjunction(data.base_infos?.length, index)}{" "}
                   </>
                 ))}
               </Box>
@@ -41,6 +58,17 @@ const CompareResultContainer = () => {
         );
       }}
     />
+  );
+};
+
+const getConjunction = (length: number, index: number) => {
+  return index < length - 1 ? (
+    <>
+      {" "}
+      <Trans i18nKey="and" />
+    </>
+  ) : (
+    ""
   );
 };
