+++ frontend/src/config/translations/en.json
@@ -145,15 +145,15 @@
   "thisReportIsNotAccurate": "The report is not accurate enough.",
   "pleaseCompleteAllSubjectQuestionnaires": "Please complete all {{subjectName}} questionnaires to make it more accurate.",
   "pleaseCompleteAllQuestionnaires": "Please complete more questionnaires to get a better accuracy.",
-  "toGetMoreAccurateInsights": "In order to get more accurate insights, fill all questionnaires.",
-  "selectAssessmentForComparison": "Select an assessment for comparison.",
+  "toGetMoreAccurateInsights": "In order to get more accurate insights, compare",
   "selectAssessment": "Select assessment",
   "changeAssessment": "Change assessment",
   "addToCompareList": "Add to compare list",
   "statusIsEvaluatedAs": "Status is evaluated as",
   "changeSelectedAssessment": "Change selected assessment",
   "assessmentsAreFilteredBy": "Assessments are filtered by",
-  "inOrderToSelectAssessmentsFromOtherProfiles": "In order to select assessments from other profiles make a new comparison.",
+  "selectAssessmentForComparison": "Select the {{value}} assessment to compare",
+  "toCompareAssessmentsOfOtherProfiles": "To compare assessments of other profiles, please create a new comparison.",
   "newComparison": "New comparison",
   "addToCompare": "Add to compare",
   "theResultOfComparing": "The result of comparing",
@@ -165,5 +165,10 @@
   "answers": "Answers",
   "generalSpecification": "General specification",
   "allOfTheSelectedAssessmentsUse": "All of the selected assessments use",
-  "whichHasNamed": "Which has {{value}} subjects named"
+  "whichHasNamed": "Which has {{value}} subjects named",
+  "onlyAssessmentsWithEvaluatedStatus": "Only assessments with evaluated status can be selected to compare.",
+  "and": "and",
+  "attributes": "Attributes",
+  "subjectAttributes": "{{subject}} attributes",
+  "editComparisonItems": "Edit comparison items"
 }
